requirements.txt
asgiref==3.2.7
Django==3.0.6
djangorestframework==3.11.0
jsonfield==3.1.0
lorun==1.0.1
mysqlclient==1.4.6
Pillow==7.1.2
PyMySQL==0.9.3
pytz==2020.1
sqlparse==0.3.1
